Update deprecation warning
The deprecation deadline has passed, so updating the language to be more accurate.

I also specified the version that was deprecated, since I actually had misunderstood the warning to be saying that "this version" was referring to the entire action itself, and that the guidance was to no longer use the action at all!  I think the new language will help prevent that kind of confusion for others.